LayoutTests/editing/deleting/merge-paragraphs-with-transparent-background.html

<!DOCTYPE html>
<html style="background: transparent; text-decoration: none">
<body>
<p id="description">This tests merging paragraphs inside a document with background: transparent in html element's inline style declaration.
WebKit should not generate a span with background-color property in such cases.</p>
<div id="container" contenteditable>
<p id="destination">hello</p>
<p id="target">world</p>
</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description(document.getElementById('description').textContent);

document.getElementById('container').focus();
var target = document.getElementById('target');
getSelection().collapse(target, 0);
document.execCommand('Delete', false, null);

Markup.dump(document.getElementById('container'));

</script>
</body>
</html>
