LayoutTests/editing/deleting/merge-paragraphs-with-transparent-background-expected.txt

This tests merging paragraphs inside a document with background: transparent in html element's inline style declaration.
CyberKit should not generate a span with background-color property in such cases.
| "
"
| <p>
|   id="destination"
|   "hello<#selection-caret>world"
| "
"
